stubs/pyproject.toml
[project]
name = "dummy_api"
version = "0.0.0"
description = "Dummy API for testing"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "prompt-toolkit>=3.0.52",
    "pyserial>=3.5",
]
